moved unnecessary exports to utils.ts
