+++ package.json
@@ -41,6 +41,7 @@
         "webpack-cli": "^3.3.10"
     },
     "files": [
-        "dist/*"
+        "dist/index.d.ts",
+        "dist/index.min.js"
     ]
 }
